fix "unknown" status
sometimes PR is listed as "unknown" mergeability status (rather than "behind")
after another PR is merged into the same base branch. I assume this is a github
API quirk when you query it too quickly after a change that impacts the PR
status. I think its safe to just skip this check for "behind", because: 1) this
code is only run when a PR is merged, 2) we only look for PRs that have the same
base branch and have the `merge-it` label. Therefore, the PRs we try to rebase
MUST be behind, obviating the need for a check.